set Authenticated APIs
This will set auth middleware so that only authenticaed/logged-in
user can access and modify the data. Special feature: After the
user gets deleted, its AccessToken can be used to access other
endpoints. Thus, additional check setup to check if the user is
present in the database or not.

Connected Author with Books and Like functionalities remained.